add updated callback funcs to event package
